requirements.txt
Flask
newspaper
pika
python-dateutil
python-jsonrpc
pymongo
redis
requests
sklearn